\name{rainfall_login}
\alias{rainfall_login}
\title{Better Rainfall Forecast for Grain Growers API connection}
\usage{
  rainfall_login(username, password)
}
\arguments{
  \item{username}{username for the API}

  \item{password}{password for the API}
}
\description{
  This function is used to use the API
}

